+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Development
 
+- Merge patch metadata and config file into one large config file.
+
 ## [0.1.0](https://github.com/berlinonline/ckan_metadata_updater/releases/tag/0.1.0)
 
 _(2022-04-05)_